+++ src/cirrus/github_tools.py
@@ -1,7 +1,6 @@
 '''
 Contains class for handling the creation of pull requests
 '''
-import itertools
 import json
 import requests
 
@@ -61,44 +60,6 @@ def create_pull_request(
     return resp_json['html_url']
 
 
-def get_tags_with_sha(owner, repo, token=None):
-    """
-    _get_tags_with_sha_
-
-    Get list of tags for a repo and return a map of
-    tag:sha
-
-    """
-    url = "https://api.github.com/repos/{owner}/{repo}/tags".format(
-        owner=owner, repo=repo
-    )
-    if token is None:
-        token = get_github_auth()[1]
-
-    headers = {
-        'Authorization': 'token %s' % token
-    }
-
-    resp = requests.get(url, headers=headers)
-    resp.raise_for_status()
-    result = { r['name']:r['commit']['sha'] for r in resp.json()}
-    return result
-
-
-def get_tags(owner, repo, token=None):
-    """
-    _get_tags_
-
-    returns a list of tags for the given repo, ordered as
-    newest first
-
-    """
-    tags_with_sha = get_tags_with_sha(owner, repo, token)
-    tags = tags_with_sha.keys()
-    tags.sort(reverse=True)
-    return tags
-
-
 def get_releases(owner, repo, token=None):
 
     url = "https://api.github.com/repos/{owner}/{repo}/releases".format(
@@ -116,134 +77,3 @@ def get_releases(owner, repo, token=None):
 
     releases = [ release for release in resp.json() ]
     return releases
-
-
-def get_commit_msgs(owner, repo, since_sha, token=None):
-    """
-    _get_commit_msgs_
-
-    Get commit message data for the repo provided since the
-    since_sha value of a commit or tag.
-
-    """
-    url = "https://api.github.com/repos/{owner}/{repo}/commits".format(
-        owner=owner, repo=repo
-    )
-
-    params = {'sha': since_sha}
-    if token is None:
-        token = get_github_auth()[1]
-
-    headers = {
-        'Authorization': 'token %s' % token
-    }
-
-    resp = requests.get(url, headers=headers, params=params)
-    resp.raise_for_status()
-    result = []
-    for commit in resp.json():
-
-        blame = None
-        if commit.get("committer"):
-            blame = commit['committer'].get('login')
-
-        if blame is None:
-            # Handle commits that weren't properly identified with a
-            # GitHub user, or None 'login' values
-            blame = commit['commit']['committer'].get('name')
-
-        result.append({
-            "committer" : blame,
-            "message":  commit['commit']['message'],
-            "date" : commit['commit']['committer']['date']
-
-        })
-    return result
-
-
-def format_commit_messages(rows):
-    """
-    _format_commit_messages_
-
-    Consume the data produced by get_commit_msgs and
-    generate a set of release notes, broken down by author
-
-    Output looks like:
-
-    - Commit History:
-    -- Author: GITHUBUSERNAME
-    --- DATETIME: COMMIT MESSAGE
-
-    """
-
-    result = [u" - Commit History:"]
-
-    for author, commits in itertools.groupby(rows, lambda x: x['committer']):
-        result.append(u" -- Author: {0}".format(author))
-        sorted_commits = sorted([ c for c in commits ], key=lambda x: x['date'], reverse=True)
-        result.extend( u' --- {0}: {1}'.format(commit['date'],commit['message']) for commit in sorted_commits)
-
-    return '\n'.join(result)
-
-
-def markdown_format(rows):
-    """
-    _format_commit_messages_
-
-    Consume the data produced by get_commit_msgs and
-    generate a set of release notes, broken down by author
-
-    Output looks like:
-
-    Commit History
-    ==============
-
-    Author: GITHUBUSERNAME
-    ----------------------
-
-    DATETIME: COMMIT MESSAGE
-
-    """
-
-    result = ['Commit History\n==============']
-
-    for author, commits in itertools.groupby(rows, lambda x: x['committer']):
-        result.append(
-            '\nAuthor: {0}\n--------'.format(author) + '-' * len(author))
-        sorted_commits = sorted(
-            [c for c in commits],
-            key=lambda x: x['date'],
-            reverse=True)
-        result.extend('\n{0}: {1}'.format(
-            commit['date'],
-            commit['message']) for commit in sorted_commits)
-
-    return '\n'.join(result)
-
-FORMATTERS = {
-    'plaintext': format_commit_messages,
-    'markdown': markdown_format,
-    }
-
-
-def build_release_notes(org, repo, since_tag, formatter):
-    """
-    Given an org, repo and tag, generate release notes for all
-    commits since that tag
-
-    """
-    tags = get_tags_with_sha(org, repo)
-    if since_tag not in tags:
-        msg = "Could not find tag {0} in {1}/{2}".format(since_tag, org, repo)
-        raise RuntimeError(msg)
-
-    sha = tags[since_tag]
-    msgs = get_commit_msgs(org, repo, sha)
-    try:
-        rel_notes = FORMATTERS[formatter](msgs)
-    except Exception as ex:
-        LOGGER.exception(ex)
-        raise RuntimeError(
-            ('Invalid release notes formatting: {0} Update cirrus.conf'
-             ' entry to use either: plaintext, markdown'.format(formatter)))
-    return rel_notes
